moved some functions back to internal
