+++ content/utils.go
@@ -12,6 +12,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
+
 package content
 
 import (
@@ -20,6 +21,7 @@ import (
 	"io"
 
 	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
+	"oras.land/oras-go/v2/internal/ioutil"
 )
 
 // ReadAll safely reads the content described by the descriptor.
@@ -37,39 +39,9 @@ func ReadAll(r io.Reader, desc ocispec.Descriptor) ([]byte, error) {
 		return nil, errors.New("digest verification failed")
 	}
 
-	if err := ensureEOF(r); err != nil {
+	if err := ioutil.EnsureEOF(r); err != nil {
 		return nil, err
 	}
 
 	return buf, nil
-}
-
-// CopyBuffer copies from src to dst through the provided buffer
-// until either EOF is reached on src, or an error occurs.
-// The copied content is verified against the size and the digest.
-func CopyBuffer(dst io.Writer, src io.Reader, buf []byte, desc ocispec.Descriptor) error {
-	// verify while copying
-	verifier := desc.Digest.Verifier()
-	lr := io.LimitReader(src, desc.Size)
-	mw := io.MultiWriter(verifier, dst)
-
-	if _, err := io.CopyBuffer(mw, lr, buf); err != nil {
-		return fmt.Errorf("copy failed: %w", err)
-	}
-
-	if !verifier.Verified() {
-		return errors.New("digest verification failed")
-	}
-
-	return ensureEOF(lr)
-}
-
-func ensureEOF(r io.Reader) error {
-	var peek [1]byte
-	_, err := io.ReadFull(r, peek[:])
-	if err != io.EOF {
-		return errors.New("trailing data")
-	}
-
-	return nil
 }
