+++ internal/ioutil/io.go
@@ -16,8 +16,12 @@ limitations under the License.
 package ioutil
 
 import (
+	"errors"
+	"fmt"
 	"io"
 	"reflect"
+
+	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
 )
 
 // CloserFunc is the basic Close method defined in io.Closer.
@@ -26,6 +30,36 @@ type CloserFunc func() error
 // Close performs close operation by the CloserFunc.
 func (fn CloserFunc) Close() error {
 	return fn()
+}
+
+// CopyBuffer copies from src to dst through the provided buffer
+// until either EOF is reached on src, or an error occurs.
+// The copied content is verified against the size and the digest.
+func CopyBuffer(dst io.Writer, src io.Reader, buf []byte, desc ocispec.Descriptor) error {
+	// verify while copying
+	verifier := desc.Digest.Verifier()
+	lr := io.LimitReader(src, desc.Size)
+	mw := io.MultiWriter(verifier, dst)
+
+	if _, err := io.CopyBuffer(mw, lr, buf); err != nil {
+		return fmt.Errorf("copy failed: %w", err)
+	}
+
+	if !verifier.Verified() {
+		return errors.New("digest verification failed")
+	}
+
+	return EnsureEOF(lr)
+}
+
+func EnsureEOF(r io.Reader) error {
+	var peek [1]byte
+	_, err := io.ReadFull(r, peek[:])
+	if err != io.EOF {
+		return errors.New("trailing data")
+	}
+
+	return nil
 }
 
 // nopCloserType is the type of `io.NopCloser()`.
